+++ changelog.md
@@ -3,8 +3,8 @@
 > The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), </br>
 > and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).</br>
 > To add a tag to the git repository do:
-> > git tag -a v5.12.0 -m "Added extras::exception ... "</br>
-> > git push origin v5.12.0
+> > git tag -a v5.13.0 -m "Improved extras::exception ... "</br>
+> > git push origin v5.13.0
 > >
 >
 > > POPULAR LINKS:
@@ -23,6 +23,10 @@ IMPORTANT:
 - MAKE SURE to tag and update the tag compare list below, (for quick code comparison purposes).<br>
 
 ## [Unreleased]
+
+#### 5.13.0 (2021-12-03)
+### Improved
+- exception handler
 
 #### 5.12.0 (2021-12-02)
 ### Improved
@@ -625,7 +629,8 @@ IMPORTANT:
 ### Removed
 - nothing
 
-[Unreleased]: https://github.com/perriera/extras/compare/v5.12.0..HEAD
+[Unreleased]: https://github.com/perriera/extras/compare/v5.13.0..HEAD
+[5.13.0]: https://github.com/perriera/extras/compare/v5.12.0...v5.13.0
 [5.12.0]: https://github.com/perriera/extras/compare/v5.11.1...v5.12.0
 [5.11.1]: https://github.com/perriera/extras/compare/v5.10.4...v5.11.1
 [5.10.4]: https://github.com/perriera/extras/compare/v5.9.2...v5.10.4